+++ xc3_model/src/material/assignments.rs
@@ -22,11 +22,14 @@ pub struct OutputAssignments {
     /// The parameter multiplied by vertex alpha to determine outline width.
     pub outline_width: Option<Value>,
 
-    /// Index into [assignments](#structfield.assignments) for the intensity map for normal mapping.
+    /// Index into [exprs](#structfield.exprs) for the intensity map for normal mapping.
     pub normal_intensity: Option<usize>,
 
-    /// Index into [assignments](#structfield.assignments) for the intensity for vValInf normal mapping.
+    /// Index into [exprs](#structfield.exprs) for the intensity for vValInf normal mapping.
     pub val_inf_intensity: Option<usize>,
+
+    /// Index into [exprs](#structfield.exprs) for the fragment discard condition.
+    pub discard_condition: Option<usize>,
 
     /// Unique values shared between all outputs.
     pub exprs: Vec<OutputExpr>,
@@ -126,6 +129,7 @@ pub(crate) fn output_assignments(
             .map(|d| assignment_value(d, parameters)),
         normal_intensity: shader.normal_intensity,
         val_inf_intensity: shader.val_inf_intensity,
+        discard_condition: shader.discard_condition,
         exprs,
         exprs_xyz,
     }
@@ -243,6 +247,7 @@ pub(crate) fn infer_assignment_from_textures(
         outline_width: None,
         normal_intensity: None,
         val_inf_intensity: None,
+        discard_condition: None,
         exprs: exprs.into_iter().collect(),
         exprs_xyz: exprs_xyz.into_iter().collect(),
     }